ready login and logout

+++ model/contacts.js
@@ -4,26 +4,31 @@ const addContact = async (body) => {
   return await Contact.create(body)
 }
 
-const listContacts = async () => {
-  return await Contact.find({})
+const listContacts = async (userId) => {
+  return await Contact.find({ owner: userId }).populate({
+    path: 'owner',
+    select: 'email',
+  })
 }
 
-const getContactById = async (contactId) => {
-  return await Contact.findById(contactId)
+const getContactById = async (contactId, userId) => {
+  return await Contact.findOne({ _id: contactId, userId }).populate({
+    path: 'owner',
+    select: 'email',
+  })
 }
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, userId) => {
   return await Contact.findByIdAndUpdate(
-    contactId,
+    { _id: contactId, owner: userId },
     { ...body },
-    {
-      new: true,
-    }
+    { new: true }
   )
 }
 
-const removeContact = async (contactId) => {
-  return await Contact.findByIdAndRemove(contactId)
+const removeContact = async (contactId, userId) => {
+  console.log(contactId)
+  return await Contact.findByIdAndRemove({ _id: contactId, owner: userId })
 }
 
 module.exports = {
